// src/enums.ts
export enum EElement {
    Earth = 'Earth',
    Fire = 'Fire',
    Light = 'Light',
    Dark = 'Dark',
    Basic = 'Basic',
    Water = 'Water'
}

export enum ERole {
    Warrior = 'Warrior',
    Support = 'Support',
    Tanker = 'Tanker',
    Ranged = 'Ranged'
}

export enum EBuff {
    PARTY_HP = '[Party] HP',
    PARTY_SKILL_ATK = '[Party] Skill Atk',
    PARTY_DEF = '[Party] Def',
    PARTY_RANGED_ATK = '[Party] Range Atk',
    SKILL_ATK = 'Skill Atk',
    WEAPON_SKILL_REGEN_SPEED = 'Weapon Skill Regen Speed',
    SHIELD_INCREASE_ON_BATTLE_START = 'Shield increase on battle start',
    HP_ON_ENEMY_KILL = 'HP Recovery on enemy kill',
    ATK_ON_ENEMY_KILL = 'Atk increase on enemy kill',
}